requirements.txt
flask
python-dotenv
pymongo[srv]
dnspython
certifi
sentence-transformers
torch
pypdf
python-docx
google-genai
